+++ src/Filters/AuthenticationFilter.php
@@ -2,6 +2,7 @@
 
 namespace Fluent\Auth\Filters;
 
+use CodeIgniter\API\ResponseTrait;
 use CodeIgniter\Filters\FilterInterface;
 use CodeIgniter\HTTP\RequestInterface;
 use CodeIgniter\HTTP\ResponseInterface;
@@ -12,12 +13,18 @@ use Fluent\Auth\Exceptions\AuthenticationException;
 
 class AuthenticationFilter implements FilterInterface
 {
+    use ResponseTrait;
+
     /** @var AuthFactoryInterface|AuthenticationInterface */
     protected $auth;
+
+    /** @var ResponseInterface */
+    protected $response;
 
     public function __construct()
     {
         $this->auth = Services::auth();
+        $this->response = Services::response();
     }
 
     /**
@@ -25,9 +32,7 @@ class AuthenticationFilter implements FilterInterface
      */
     public function before(RequestInterface $request, $arguments = null)
     {
-        $this->authenticate($request, $arguments);
-
-        return $request;
+        return $this->authenticate($request, $arguments);
     }
 
     /**
@@ -57,7 +62,7 @@ class AuthenticationFilter implements FilterInterface
             }
         }
 
-        $this->unauthenticated($request, $guards);
+        return $this->unauthenticated($request, $guards);
     }
 
      /**
@@ -70,6 +75,10 @@ class AuthenticationFilter implements FilterInterface
       */
     protected function unauthenticated($request, $guards)
     {
+        if ($request->isAJAX()) {
+            return $this->fail('Unauthenticated.', ResponseInterface::HTTP_UNAUTHORIZED);
+        }
+
         throw new AuthenticationException(
             'Unauthenticated.',
             $guards,